+++ packages/subscriptions/src/getSubscriptionStatus.ts
@@ -1,8 +1,9 @@
 import type { SpaceSubscriptionTier } from '@charmverse/core/prisma';
 import { prisma } from '@charmverse/core/prisma-client';
 import { getSpaceTokenBalance } from '@packages/spaces/getSpaceTokenBalance';
-import { tierConfig } from '@packages/subscriptions/constants';
 import { DateTime } from 'luxon';
+
+import { getExpiresAt } from './getExpiresAt';
 
 export type SpaceSubscriptionStatus = {
   tier: SpaceSubscriptionTier;
@@ -55,17 +56,4 @@ export async function getSubscriptionStatus(spaceId: string) {
     expiresAt,
     isReadonlyNextMonth
   };
-}
-
-// undefined balance means it is loading still
-function getExpiresAt(tier: SpaceSubscriptionTier | null, spaceTokenBalance: number) {
-  if (!tier) {
-    return null;
-  }
-  const tierPrice = tierConfig[tier].tokenPrice;
-  if (tierPrice > 0) {
-    const monthsLeft = Math.floor(spaceTokenBalance / tierPrice);
-    return DateTime.utc().endOf('month').plus({ months: monthsLeft }).endOf('month');
-  }
-  return null;
 }
